refactor(frontend): decompose ChatContext and extract modular UI components
- Extract domain types and interfaces from ChatContext.tsx into context/types.ts
- Extract pure mappers, serializers, and helpers into context/chatMappers.ts
- Add dedicated unit test suite tests/chatMappers.test.ts (16 tests)
- Extract MessageRow and TypingIndicator components from Chatroom.tsx
- Extract RoomItem and SectionLabel components from ChatList.tsx
- Clean up unused variables and imports across frontend code and tests
- Maintain full backwards compatibility via re-exports from ChatContext
